TEXLoader - HKLUT perfromance measurement added; minor refactor

+++ ogre/src/loaders/TEXLoader.h
@@ -38,8 +38,8 @@ private:
     void doMSB(size_t x, size_t y, const uint8_t* src, size_t stride, const LUTs& luts, std::vector<PixelSigned>& res) const;
     Pixel getPixel(int x, int y, const uint8_t* src, size_t stride) const;
 
-    std::vector<int16_t> getLUTValsLSB(const std::vector<int8_t>& lut, const Pixel& valA, const Pixel& valB) const;
-    std::vector<int16_t> getLUTValsMSB(const std::vector<int8_t>& lut, const Pixel& valA, const Pixel& valB, const Pixel& valC) const;
+    std::vector<int8_t> getLUTValsLSB(const std::vector<int8_t>& lut, const Pixel& valA, const Pixel& valB) const;
+    std::vector<int8_t> getLUTValsMSB(const std::vector<int8_t>& lut, const Pixel& valA, const Pixel& valB, const Pixel& valC) const;
     void rotateBack(std::vector<int16_t>& res) const;
 
     std::vector<uint8_t> toRGB(Ogre::Image& img) const;
